backend/requirements.txt
fastapi
uvicorn
openai-whisper
transformers
gtts
language-tool-python